+++ packages/@expo/cli/src/export/exportApp.ts
@@ -61,7 +61,6 @@ export async function exportAppAsync(
     {
       platforms,
       dev,
-      useDevServer: true,
       // TODO: Disable source map generation if we aren't outputting them.
     }
   );