+++ packages/@expo/cli/src/api/graphql/types/App.ts
@@ -1,0 +1,11 @@
+import gql from 'graphql-tag';
+
+export const AppFragmentNode = gql`
+  fragment AppFragment on App {
+    id
+    scopeKey
+    ownerAccount {
+      id
+    }
+  }
+`;