+++ packages/@expo/cli/src/start/server/middleware/ManifestMiddleware.ts
@@ -61,8 +61,6 @@ export interface HostInfo {
 
 /** Parsed values from the supported request headers. */
 export interface ManifestRequestInfo {
-  /** Should return the signed manifest. */
-  acceptSignature: boolean;
   /** Platform to serve. */
   platform: RuntimePlatform;
   /** Requested host name. */
@@ -163,18 +161,6 @@ export abstract class ManifestMiddleware<
 
   /** Parse request headers into options. */
   public abstract getParsedHeaders(req: ServerRequest): TManifestRequestInfo;
-
-  /**
-   * This header is specified as a string "true" or "false", in one of two headers:
-   * - exponent-accept-signature
-   * - expo-accept-signature
-   */
-  protected getLegacyAcceptSignatureHeader(req: ServerRequest): boolean {
-    return (
-      req.headers['exponent-accept-signature'] === 'true' ||
-      req.headers['expo-accept-signature'] === 'true'
-    );
-  }
 
   /** Store device IDs that were sent in the request headers. */
   private async saveDevicesAsync(req: ServerRequest) {