+++ packages/@expo/cli/CHANGELOG.md
@@ -6,9 +6,12 @@
 
 ### 🎉 New features
 
+- Implement new package manager API in CLI. ([#19343](https://github.com/expo/expo/pull/19343) by [@byCedric](https://github.com/byCedric))
+
 ### 🐛 Bug fixes
 
 - Fix log format when modifying `tsconfig.json`. ([#21166](https://github.com/expo/expo/pull/21166) by [@EvanBacon](https://github.com/EvanBacon))
+- Fix `devDependencies` when running `npx expo install --fix`. ([#19344](https://github.com/expo/expo/pull/19344) by [@byCedric](https://github.com/byCedric))
 
 ### 💡 Others