+++ packages/@expo/cli/CHANGELOG.md
@@ -25,6 +25,7 @@
 ### 🐛 Bug fixes
 
 - Respond to `Debugger.getScriptSource` CDP messages when using lan or tunnel. ([#21825](https://github.com/expo/expo/pull/21825) by [@byCedric](https://github.com/byCedric))
+- Fix main field resolution for metro web. ([#21939](https://github.com/expo/expo/pull/21939) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 💡 Others